Containerized a simple web application
Built a simple web application with FastAPI, still need to make it REST.
Defined a Dockerfile.
Defined a docker-compose.yml
Converted the command line application from `main.py` to `main-cli.py`.
Updated requirements in favor of the web application.
