+++ docker-compose.yml
@@ -1,0 +1,8 @@
+version: '2.2'
+services:
+  novel-crawlers:
+    container_name: novel-crawler
+    image: nov-craw
+    ports:
+      - "10000:10000"
+    restart: unless-stopped
