+++ requirements.txt
@@ -1,2 +1,18 @@
-bs4
-requests
+anyio==3.4.0
+asgiref==3.4.1
+beautifulsoup4==4.9.3
+bs4==0.0.1
+certifi==2020.12.5
+chardet==4.0.0
+click==8.0.3
+fastapi==0.70.0
+h11==0.12.0
+idna==2.10
+pydantic==1.8.2
+requests==2.25.1
+sniffio==1.2.0
+soupsieve==2.2.1
+starlette==0.16.0
+typing-extensions==4.0.1
+urllib3==1.26.4
+uvicorn==0.15.0
